+++ src/ScrippsGenomeAdviserUI/heapSortAlgorithm.java
@@ -88,10 +88,10 @@ public class heapSortAlgorithm extends JFrame implements Runnable {
      //  ArrayforSort = new double[dataLines][2];
         
        
-                    double y = (double) dataLines;
+                    double y = dataLines;
                     double ing  = 100000.0;
-                    double x = (double) ing/dataLines;
-                    double perc = (double) (x*100);
+                    double x = ing/dataLines;
+                    double perc = x*100;
 
          
         //Extracting the row number to sort by
@@ -193,7 +193,7 @@ public class heapSortAlgorithm extends JFrame implements Runnable {
             for (int i = 0; i < ac.length(); i++){
             char ch = ac.charAt(i);  
             //here is the trick from char to ASCII
-            int jh = (int) ch;
+            int jh = ch;
             String n1 = Integer.toString(jh); 
             n = n + n1;
             }
