+++ src/org/scripps/utils/CnvFilterFunctions.java
@@ -94,7 +94,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculations
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -181,7 +181,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculations
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -424,7 +424,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculations
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -493,7 +493,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculation
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -587,7 +587,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculation
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -714,7 +714,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculations
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -822,7 +822,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculations
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -936,7 +936,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculations
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -1035,7 +1035,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculations
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -1153,7 +1153,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculations
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -1262,7 +1262,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculations
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -1438,7 +1438,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculations
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -1542,7 +1542,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculations
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -1665,7 +1665,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculations
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
@@ -2210,7 +2210,7 @@ public class CnvFilterFunctions implements Runnable
 		// progress bar calculations
 		int lineCount = 0;
 		int linesLeft = end;
-		int threePerc = (int) (linesLeft / 100) * 3;
+		int threePerc = linesLeft / 100 * 3;
 		if (threePerc < 1)
 		{
 			threePerc = 1;
