+++ src/org/scripps/utils/CnvHeapSortAlgorithm.java
@@ -89,10 +89,10 @@ public class CnvHeapSortAlgorithm extends JFrame implements Runnable
 		// ArrayforSort = null;
 		// ArrayforSort = new double[dataLines][2];
 
-		double y = (double) dataLines;
+		double y = dataLines;
 		double ing = 100000.0;
-		double x = (double) ing / dataLines;
-		double perc = (double) (x * 100);
+		double x = ing / dataLines;
+		double perc = x * 100;
 
 		// Extracting the row number to sort by
 
@@ -222,7 +222,7 @@ public class CnvHeapSortAlgorithm extends JFrame implements Runnable
 						{
 							char ch = ac.charAt(i);
 							// here is the trick from char to ASCII
-							int jh = (int) ch;
+							int jh = ch;
 							String n1 = Integer.toString(jh);
 							n = n + n1;
 						}
