+++ src/org/scripps/utils/CnvStatistics.java
@@ -501,7 +501,7 @@ public class CnvStatistics implements Runnable
 			// progress bar calculations
 			int lineCount = 0;
 			int linesLeft = end;
-			int threePerc = (int) (linesLeft / 100) * 3;
+			int threePerc = linesLeft / 100 * 3;
 			if (threePerc < 1)
 			{
 				threePerc = 1;
